Add/move tests to new module
